client/requirements.txt
black==24.8.0
discord.py=2.4.0
flake8==7.0.0
isort==5.12.0
pre-commit==3.8
pre-commit-hooks=4.6.0
python-dotenv=1.0.1
